requirements.txt
appdirs==1.4.0
Django==1.10.5
django-cors-headers==2.0.2
djangorestframework==3.5.3
drfdocs==0.0.11
gunicorn==19.7.0
Markdown==2.6.8
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
